+++ sphinx/source/changelog.rst
@@ -23,6 +23,7 @@ Changelog
   in the ``TraceLine`` configuration parameter
 * Added support to :ref:`bin2sna.py` and :ref:`snapmod.py` for setting the
   MEMPTR register in SZX snapshots
+* Added support to the :ref:`AUDIO` macro for keyword arguments
 * Added support to :ref:`rzxplay.py` for ignoring any snapshots after the first
   one (which is useful for playing some RZX files created by the Fuse emulator)
 * Fixed the bug in CSimulator that restricts the ``timeout`` simulated load